# S4C_Admin/temp_swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Study4Clone API",
    "description": "IELTS Mock Test Backend API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Exams": {
      "get": {
        "tags": [
          "Exams"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ExamStatus"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ExamType"
            }
          },
          {
            "name": "skill",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SkillType"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamListDtoPagedList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamListDtoPagedList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamListDtoPagedList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Exams"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Exams/{id}": {
      "get": {
        "tags": [
          "Exams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Exams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExamRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExamRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExamRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Exams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Exams/{id}/full": {
      "get": {
        "tags": [
          "Exams"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Exam"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exam"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exam"
                }
              }
            }
          }
        }
      }
    },
    "/api/Exams/by-slug/{slug}": {
      "get": {
        "tags": [
          "Exams"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Skills/exam/{examId}": {
      "get": {
        "tags": [
          "Skills"
        ],
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExamSkillListDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExamSkillListDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExamSkillListDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Skills"
        ],
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamSkillRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamSkillRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamSkillRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamSkillListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamSkillListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamSkillListDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttemptStatus": {
        "enum": [
          "InProgress",
          "Pending",
          "Completed",
          "Abandoned"
        ],
        "type": "string"
      },
...
      "CreateExamRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ExamType"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CreateExamSkillRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "skill": {
            "$ref": "#/components/schemas/SkillType"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Exam": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ExamType"
          },
          "status": {
            "$ref": "#/components/schemas/ExamStatus"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamSkill"
            },
            "nullable": true
          },
          "attempts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestAttempt"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExamDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ExamType"
          },
          "status": {
            "$ref": "#/components/schemas/ExamStatus"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ExamListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ExamType"
          },
          "status": {
            "$ref": "#/components/schemas/ExamStatus"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "availableSkills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "targetSkill": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExamListDtoPagedList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamListDto"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "readOnly": true
          },
          "hasNextPage": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ExamSection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "skillId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "textContent": {
            "type": "string",
            "nullable": true
          },
          "audioUrl": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "transcript": {
            "type": "string",
            "nullable": true
          },
          "skill": {
            "$ref": "#/components/schemas/ExamSkill"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionGroup"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ExamSkill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "examId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "skill": {
            "$ref": "#/components/schemas/SkillType"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "exam": {
            "$ref": "#/components/schemas/Exam"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamSection"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ExamSkillListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "skill": {
            "$ref": "#/components/schemas/SkillType"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "sectionCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ExamStatus": {
        "enum": [
          "Draft",
          "Published",
          "Archived"
        ],
        "type": "string"
      },
      "ExamType": {
        "enum": [
          "IeltsAcademic",
          "IeltsGeneral",
          "Toeic",
          "Other"
        ],
        "type": "string"
      },
...
      "QuestionGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sectionId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "instruction": {
            "type": "string",
            "nullable": true
          },
          "questionType": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "textContent": {
            "type": "string",
            "nullable": true
          },
          "audioUrl": {
            "type": "string",
            "nullable": true
          },
          "section": {
            "$ref": "#/components/schemas/ExamSection"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "QuestionType": {
        "enum": [
          "MultipleChoice",
          "FillInTheBlank",
          "TrueFalseNotGiven",
          "MatchingHeadings",
          "DropDown",
          "Essay",
          "SpeakingRecording"
        ],
        "type": "string"
      },
...
      "SkillType": {
        "enum": [
          "Listening",
          "Reading",
          "Writing",
          "Speaking"
        ],
        "type": "string"
      },
...
      "TestAttempt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "examId": {
            "type": "string",
            "format": "uuid"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/AttemptStatus"
          },
          "timeRemaining": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "selectedSkills": {
            "type": "string",
            "nullable": true
          },
          "overallScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "listeningScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "readingScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "writingScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "speakingScore": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "correctQuestionCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalQuestionCount": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "exam": {
            "$ref": "#/components/schemas/Exam"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAnswer"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateExamRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ExamType"
          },
          "status": {
            "$ref": "#/components/schemas/ExamStatus"
          },
          "duration": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "attempts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestAttempt"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "attemptId": {
            "type": "string",
            "format": "uuid"
          },
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "audioUrl": {
            "type": "string",
            "nullable": true
          },
          "textContent": {
            "type": "string",
            "nullable": true
          },
          "selectedOptions": {
            "type": "string",
            "nullable": true
          },
          "isCorrect": {
            "type": "boolean",
            "nullable": true
          },
          "scoreEarned": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "feedback": {
            "type": "string",
            "nullable": true
          },
          "attempt": {
            "$ref": "#/components/schemas/TestAttempt"
          },
          "question": {
            "$ref": "#/components/schemas/Question"
          }
        },
        "additionalProperties": false
      },
      "UserRole": {
        "enum": [
          "User",
          "Admin",
          "ContentCreator"
        ],
        "type": "string"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}